Minor structure update
- Moved savefile management to seperate file
- Tested the savefile functions
- Adjusted other files spelling mistake (copied from template)
