+++ gamedata/data/mainscreen/scripts/peasant.lua
@@ -1,5 +1,5 @@
 --[[
-	Testclass for "Filthy Peasants" peasant
+	Peasantclass for "Filthy Peasants"
 	©LazyProggers
 ]]--
