+++ gamedata/data/templates/scripts/template_luascript.lua
@@ -1,5 +1,5 @@
 --[[
-	Testclass for "Filthy Peasants" peasant
+	Testclass for "Filthy Peasants"
 	©LazyProggers
 ]]--
